requirements.txt
databricks-test~=0.0.4
delta-spark~=1.1.0
findspark~=1.4.2
jproperties~=2.1.1
pyspark~=3.2.1
pytest~=6.2.5
pyspark-test~=0.2.0
setuptools~=45.2.0
wheel~=0.37.1
kafka-python~=2.0.2